ui界面 10 期 清除所有的编译警告

+++ LurenjiaEngine/LurenjiaEngine/Engine/Component/Mesh/Core/MeshComponent.cpp
@@ -22,7 +22,7 @@ void CMeshComponent::SetSubMaterials(const int& index, CMaterial* InMaterial)
 
 UINT CMeshComponent::GetMaterialsCount() const
 {
-	return Materials.size();
+	return static_cast<UINT>(Materials.size());
 }
 
 const vector<CMaterial*>* CMeshComponent::GetMaterials() const
